agents/official/sequential-workflow/pyproject.toml
[project]
name = "sequential-workflow"
version = "0.1.0"
authors = [
    { name = "IBM Corp." },
]
requires-python = ">=3.13,<4"
dependencies = [
    "beeai-sdk",
    "pyyaml>=6.0.2",
]

[project.scripts]
server = "sequential_workflow.agent:run"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.uv.sources]
beeai-sdk = { path = "../../../apps/beeai-sdk", editable = true }

[tool.ruff]
line-length = 120
